+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-* Fixed bug in Discord message not being posted
-* Added version check at startup
-* Added tests for the filters
+* **Important** Requires a new `autoCoins.json` file!
+* Added excludeList. Coins added to this list will not be quarantined.
+* Added a Google Sheet whitelist